fix: add class for uploaded offer photos, reset form now reset marker on map also, add style for error container
