+++ js/main.js
@@ -1,12 +1,15 @@
+import './form.js';
 import { showError } from './util.js';
 import './slider.js';
 import { setUserFromSubmit } from './form-validation.js';
 import { createMarker } from './map.js';
 import { successPopup, errorPopup } from './popup.js';
 import { getData } from './api.js';
+const OFFERS_COUNT = 10;
+
 
 getData((offers) => {
-  offers.slice(0, 10).forEach((point) => createMarker(point));
+  offers.slice(0, OFFERS_COUNT).forEach((point) => createMarker(point));
 }, () => showError('Не удалось получить данные. Попробуйте ещё раз'));
 
 setUserFromSubmit(successPopup, errorPopup);
